remove MapController, MapView, rename MapContainer, filter closes on region select
