+++ epistimis-next/src/Components/App/App.css
@@ -1,4 +1,4 @@
-#map_controller{
+#header {
     display: flex;
     flex-direction: row;
     align-items: center;
@@ -9,3 +9,9 @@
     padding-top: 25px;
     border-bottom: 1px solid var(--background-secondary)
 }
+/* 
+#map_container {
+    position: relative;
+    z-index: 1000;
+    background-color: transparent;
+} */
