--- epistimis-next/src/Components/MapView/MapView.css
@@ -1,3 +1,0 @@
-#map_view {
-    /* padding-top: 20px; */
-}
